Infer "Origin" header from WebSocket URI
I can't connect to my ActionCable endpoint without removing the default
same-origin restriction.  This seems like a common case.

Send an Origin header that matches the WebSocket URI so that
ActionCableClient can connect to a typical Rails app out of the box.

This relies on the latest "master" version of
mwylde/em-websocket-client, which adds the :origin option.