Allow passing params when subscribing to a channel
In Rails, the ActionCable client JS library allows you to pass
additional parameters with the channel name.

    App.cable.subscriptions.create "AppearanceChannel"

vs.

    App.cable.subscriptions.create {
      channel: "ChatChannel",
      room: "Best Room"
    }

The extra parameters (like "room") seem to get incorporated into the
subscription identifier.

Add this option to ActionCableClient: allow subscribing to a simple
channel as a string or with a Hash identifier that contains a :channel
key.